Cardano.Spec.Chain.STS.Block

Documentation

data BlockHeader

Constructors

BlockHeader

Fields

_bhPrevHash ∷ !Hash

Hash of the previous block header, or genesisHash in case of the first block in a chain.

_bhSlot ∷ !Slot

Absolute slot for which the block was generated.

_bhIssuer ∷ !VKey

Block issuer.

_bhSig ∷ !(Sig Hash)

Part of the block header which must be signed.

_bhUtxoHash ∷ !Hash

UTxO hash

_bhDlgHash ∷ !Hash

Delegation hash

_bhUpdHash ∷ !Hash

Update payload hash

bhUtxoHash ∷ Lens' BlockHeader Hash

bhUpdHash ∷ Lens' BlockHeader Hash

bhSlot ∷ Lens' BlockHeader Slot

bhSig ∷ Lens' BlockHeader (Sig Hash)

bhPrevHash ∷ Lens' BlockHeader Hash

bhIssuer ∷ Lens' BlockHeader VKey

bhDlgHash ∷ Lens' BlockHeader Hash

data BlockBody

Constructors

BlockBody

Fields

_bDCerts ∷ ![DCert]

Delegation certificates

_bUtxo ∷ ![Tx]

UTxO payload

_bUpdProp ∷ !(Maybe UProp)

Update proposal payload

_bUpdVotes ∷ ![Vote]

Update votes payload

_bProtVer ∷ !ProtVer

Protocol version

bUtxo ∷ Lens' BlockBody [Tx]

bUpdVotes ∷ Lens' BlockBody [Vote]

bUpdProp ∷ Lens' BlockBody (Maybe UProp)

bProtVer ∷ Lens' BlockBody ProtVer

bDCerts ∷ Lens' BlockBody [DCert]

data Block

A block in the chain. The specification only models regular blocks since epoch boundary blocks will be largely ignored in the Byron-Shelley bridge.

Constructors

Block

Fields

_bHeader ∷ BlockHeader

_bBody ∷ BlockBody

bHeader ∷ Lens' Block BlockHeader

bBody ∷ Lens' Block BlockBody

mkBlock

Arguments

∷ Hash
Hash of the previous block

→ Slot
Current slot

→ VKey
Issuer

→ ProtVer
Protocol version

→ [DCert]
Delegation certificates

→ Maybe UProp
Update proposal

→ [Vote]
Votes on update proposals

→ [Tx]
UTxO payload

→ Block

genesisHash ∷ Hash

Dummy genesis hash.

bEndorsment ∷ Block → (ProtVer, VKey)

Protocol version endorsment

bSlot ∷ Block → Slot

Slot the block is published in

bUpdPayload ∷ Block → (Maybe UProp, [Vote])

Block update payload

bSize ∷ Block → Natural

Compute the abstract size (in words) that a block takes.

bBodySize ∷ BlockBody → Natural

Compute the abstract size (in words) that a block body occupies.

bHeaderSize ∷ BlockHeader → Natural

Compute the abstract size (in words) that a block header occupies.

hashHeader ∷ BlockHeader → Hash

Computes the hash of a header.

bhToSign ∷ BlockHeader → Hash

Computes the hash of the header.

bhHash ∷ BlockHeader → Hash

bIsEBB ∷ Block → Bool

Checks if a block is an epoch boundary block.

The function always returns False because tests will be performed only against chains without EBBs.

updateBody ∷ Block → (BlockBody → BlockBody) → Block

Update a field of the block body, recomputing the hashes to get a valid block.

data BlockStats

Constructors

BlockStats

Fields

blockStatsUtxo ∷ Word

Number of regular transactions

blockStatsDCerts ∷ Word

Number of delegation certificates

blockStatsUpdVotes ∷ Word

Number of update votes

blockStatsUpdProp ∷ Word

Number of update proposals

For a single block this will be 0 or 1.

blockStats ∷ Block → BlockStats

Count number of transactions in the block

Returns the number of

Regular transactions

Delegation certificates

Update votes

Update proposals (0 or 1)

chainBlockStats ∷ [BlockStats] → Maybe (BlockStats, BlockStats, BlockStats)

Block stats for an entire chain

Computes minimum, maximum, and average values.

Returns Nothing for the empty chain.